pyproject.toml
[tool.poetry]
name = "bam-api-connect"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
poetry-dotenv = "^0.4.0"
genai = "^2.1.0"
ibm-generative-ai = "^2.1.1"
pandas = "^2.2.0"
chardet = "^5.2.0"
scikit-learn = "^1.4.1.post1"
datasets = "^2.17.1"
plotly = "^5.19.0"
seaborn = "^0.13.2"
poetry-plugin-dotenv = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
